Reorganized file structure and added myself as an author.
